# Natural white space