--- a/src/trix/controllers/input_controller.coffee
+++ b/src/trix/controllers/input_controller.coffee
@@ -61,7 +61,7 @@ class Trix.InputController extends Trix.BasicObject
     unless @inputSummary.composing
       @handleInput ->
         unless @mutationIsExpected(mutationSummary)
-          @responder?.replaceHTML(@element.innerHTML)
+          @responder?.replaceHTML(@element.innerHTML, referenceElement: @element)
 
         @resetInputSummary()
         @requestRender()