--- a/src/trix/views/text_view.coffee
+++ b/src/trix/views/text_view.coffee
@@ -4,13 +4,12 @@ class Trix.TextView extends Trix.ObjectView
   constructor: ->
     super
     @text = @object
-    {@textConfig} = @options
 
   createNodes: ->
     nodes = []
     pieces = (piece for piece in @text.getPieces() when not piece.hasAttribute("blockBreak"))
     objects = Trix.ObjectGroup.groupObjects(pieces)
     for object in objects
-      view = @findOrCreateCachedChildView(Trix.PieceView, object, {@textConfig})
+      view = @findOrCreateCachedChildView(Trix.PieceView, object)
       nodes.push(view.getNodes()...)
     nodes