--- a/test/src/unit/html_parser_test.coffee
+++ b/test/src/unit/html_parser_test.coffee
@@ -2,15 +2,19 @@
 
 testGroup "Trix.HTMLParser", ->
   eachFixture (name, {html, serializedHTML, document}) ->
-    test name, ->
-      parsedDocument = Trix.HTMLParser.parse(html).getDocument()
-      assert.documentHTMLEqual parsedDocument.copyUsingObjectsFromDocument(document), html
+    test name, (done) ->
+      getReferenceElement (referenceElement) ->
+        parsedDocument = Trix.HTMLParser.parse(html, {referenceElement}).getDocument()
+        assert.documentHTMLEqual parsedDocument.copyUsingObjectsFromDocument(document), html
+        done()
 
   eachFixture (name, {html, serializedHTML, document}) ->
     if serializedHTML?
-      test "#{name} (serialized)", ->
-        parsedDocument = Trix.HTMLParser.parse(serializedHTML).getDocument()
-        assert.documentHTMLEqual parsedDocument.copyUsingObjectsFromDocument(document), html
+      test "#{name} (serialized)", (done) ->
+        getReferenceElement (referenceElement) ->
+          parsedDocument = Trix.HTMLParser.parse(serializedHTML, {referenceElement}).getDocument()
+          assert.documentHTMLEqual parsedDocument.copyUsingObjectsFromDocument(document), html
+          done()
 
   test "parses absolute image URLs", ->
     src = "#{getOrigin()}/test_helpers/fixtures/logo.png"
@@ -30,7 +34,7 @@ testGroup "Trix.HTMLParser", ->
 
   test "parses unfamiliar html", ->
     html = """<meta charset="UTF-8"><span style="font-style: italic">abc</span><span>d</span><section style="margin:0"><blink>123</blink><a href="http://example.com">45<b>6</b></a>x<br />y</section><p style="margin:0">9</p>"""
-    expectedHTML = """<div><!--block--><em>abc</em>d</div><div><!--block-->123<a href="http://example.com">45<strong>6</strong></a>x<br>y</div><div><!--block-->9</div>"""
+    expectedHTML = """<div><!--block--><em>abc</em>d</div><div><!--block-->123<a href="http://example.com">45<strong>6</strong></a>x\ny</div><div><!--block-->9</div>"""
     assert.documentHTMLEqual Trix.HTMLParser.parse(html).getDocument(), expectedHTML
 
   test "ignores leading whitespace before <meta> tag", ->
@@ -60,12 +64,12 @@ testGroup "Trix.HTMLParser", ->
 
   test "translates tables into plain text", ->
     html = """<table><tr><td>a</td><td>b</td></tr><tr><td>1</td><td><p>2</p></td></tr><table>"""
-    expectedHTML = """<div><!--block-->a | b<br>1 | 2</div>"""
+    expectedHTML = """<div><!--block-->a | b\n1 | 2</div>"""
     assert.documentHTMLEqual Trix.HTMLParser.parse(html).getDocument(), expectedHTML
 
   test "translates block element margins to newlines", ->
     html = """<p style="margin: 0 0 1em 0">a</p><p style="margin: 0">b</p><article style="margin: 1em 0 0 0">c</article>"""
-    expectedHTML = """<div><!--block-->a<br><br></div><div><!--block-->b</div><div><!--block--><br>c</div>"""
+    expectedHTML = """<div><!--block-->a\n<br></div><div><!--block-->b</div><div><!--block-->\nc</div>"""
     document = Trix.HTMLParser.parse(html).getDocument()
     assert.documentHTMLEqual document, expectedHTML
 
@@ -86,3 +90,18 @@ testGroup "Trix.HTMLParser", ->
 getOrigin = ->
   {protocol, hostname, port} = window.location
   "#{protocol}//#{hostname}#{if port then ":#{port}" else ""}"
+
+getReferenceElement = (callback) ->
+  if element = document.getElementById("reference_element")
+    callback(element)
+  else
+    element = document.createElement("trix-editor")
+    element.id = "reference_element"
+    element.addEventListener "trix-initialize", -> callback(element)
+
+    container = document.createElement("div")
+    container.style.position = "absolute"
+    container.style.left = "-9999px"
+
+    container.appendChild(element)
+    document.body.appendChild(container)