This change uses `white-space: pre-wrap;` for rendering newlines and consecutive spaces, and fixes #152. It's fundamentally simpler since we no longer have to render alternating `&nbsp;` to display consecutive spaces or split strings on `\n` and render `<br>`s. Less nodes are required for rendering, which yields a nice performance boost too.

---

Before and after using this document
<img width="177" alt="screen shot 2016-02-26 at 3 06 55 pm" src="https://cloud.githubusercontent.com/assets/5355/13364318/2b8180f4-dc9c-11e5-995f-0038fa3b3af7.png">

Rendered HTML *before*:
```html
<div><strong>hello</strong>&nbsp;world<br><br><em>how</em>&nbsp;&nbsp;<del>are</del>&nbsp; &nbsp;you?<br><br></div>
```

Rendered HTML *after*:
```html
<div><strong>hello</strong> world

<em>how</em>  <del>are</del>   you?
<br></div>
```

Render time *before*:
```
Rendered 10000 times
  Average: 0.131 ms
  Total: 1305.215 ms
```

Render time *after*:
```
Rendered 10000 times
  Average: 0.071 ms
  Total: 707.095 ms
```
(benchmarking setup: https://gist.github.com/javan/61544cded83763660693)

---

A potential downside is the HTML is less portable and requires its container to be styled with `white-space: pre-wrap;`, which I added to `trix.css`. I think the overall simplicity is worth that tradeoff.

/cc @sstephenson 

Trix build of this branch: [trix.zip](https://github.com/basecamp/trix/files/161929/trix.zip)