@chrisvfritz 2016-03-07
I can confirm this fixes #199. :smiley: